abstracted out settings utilities into a separate file
